+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Add `AuthenticationClass::resolve` helper function ([#432]).
 
-[#432]: https://github.com/stackabletech/operator-rs/pull/408
+[#432]: https://github.com/stackabletech/operator-rs/pull/432
 
 ## [0.22.0] - 2022-07-05